add frame registry
