+++ cmd/gnoserve/app.go
@@ -398,12 +398,7 @@ func (ds *App) RunServer(ctx context.Context, term *rawterm.RawTerm) error {
 	}
 
 	{ // Setup polling to trigger events on/off chain
-		pollRpc := ds.devNode.GetRemoteAddress()
-		if ds.proxy != nil {
-			// REVIEW: does this make sense to be here?
-			pollRpc = ds.proxy.TargetAddress() // update remote address with proxy target address
-		}
-		setupPolling(ctx, ds.logger.WithGroup("GnoPoll"), pollRpc)
+		// setupPolling(ctx, ds.logger.WithGroup("GnoPoll"), pollRpc)
 	}
 
 	if ds.cfg.interactive {
